api/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
python-dotenv
langchain
langchain-openai
slowapi
supabase
#python-jose[cryptography]
pyjwt[crypto] # Use pyjwt which has better support for JWKS
requests